Add norec build tag (#1099)
* Add norec build tag

* Add norec to recommendation
