+++ service/recommend/recommend.go
@@ -1,3 +1,5 @@
+//go:build !norec
+
 package recommend
 
 import (
@@ -6,35 +8,14 @@ import (
 	"sync"
 	"time"
 
-	"github.com/gin-gonic/gin"
 	db "github.com/mikeydub/go-gallery/db/gen/coredb"
 	"github.com/mikeydub/go-gallery/service/logger"
 	"github.com/mikeydub/go-gallery/service/persist"
-	"github.com/mikeydub/go-gallery/util"
 )
-
-const contextKey = "recommend.instance"
 
 var metadataKey = struct{}{}
 
 var currentGraph sync.Map
-
-func AddTo(c *gin.Context, r *Recommender) {
-	c.Set(contextKey, r)
-}
-
-func For(ctx context.Context) *Recommender {
-	gc := util.MustGetGinContext(ctx)
-	return gc.Value(contextKey).(*Recommender)
-}
-
-type Recommender struct {
-	RecommendFunc  func(context.Context, persist.DBID, []db.Follow) ([]persist.DBID, error)
-	LoadFunc       func(context.Context)
-	SaveResultFunc func(ctx context.Context, userID persist.DBID, recommendedIDs []persist.DBID) error
-	BootstrapFunc  func(ctx context.Context) ([]persist.DBID, error)
-	saveCh         chan saveMsg
-}
 
 func NewRecommender(queries *db.Queries) *Recommender {
 	r := &Recommender{saveCh: make(chan saveMsg)}
@@ -158,11 +139,6 @@ func (r *Recommender) Loop(ctx context.Context, ticker *time.Ticker) {
 			}
 		}
 	}()
-}
-
-type saveMsg struct {
-	nodeID         persist.DBID
-	recommendedIDs []persist.DBID
 }
 
 func (r *Recommender) readNeighbors(ctx context.Context, nodeID persist.DBID) []persist.DBID {
